+++ CHANGES.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
++ core: Remove WidgetPlus in favor of RelmWidgetExt
 + core: Add convenience getter-methods to Controller
 + core: `add_action` of `RelmActionGroup` now takes a reference to a `RelmAction` as a parameter
 + macros: `parse_with_path`, `update_stream`, `inject_view_code` and `generate_tokens` take references for some of their parameters